+++ net/oic/include/oic/messaging/coap/constants.h
