+++ net/oic/include/oic/messaging/coap/observe.h
