+++ net/oic/include/oic/messaging/coap/transactions.h
