+++ net/oic/include/oic/port/mynewt/config.h
@@ -9,6 +9,7 @@ extern "C" {
 /* Time resolution */
 #include <inttypes.h>
 #include <syscfg/syscfg.h>
+
 #include <os/os.h>
 #include <log/log.h>
