+++ net/oic/include/oic/port/oc_random.h
