+++ net/oic/src/api/oc_rep.c
@@ -15,15 +15,20 @@
 */
 
 #include <stddef.h>
+
 #include <os/os_mempool.h>
 
-#include "oic/oc_rep.h"
-#include "oic/oc_log.h"
-#include "config.h"
-#include "port/oc_assert.h"
-#include "api/oc_priv.h"
+#include <syscfg/syscfg.h>
+
 #include <tinycbor/cbor_mbuf_writer.h>
 #include <tinycbor/cbor_mbuf_reader.h>
+
+#include "oic/port/mynewt/config.h"
+#include "oic/oc_rep.h"
+#include "oic/oc_log.h"
+#include "oic/port/mynewt/config.h"
+#include "port/oc_assert.h"
+#include "api/oc_priv.h"
 
 #ifdef OC_CLIENT
 static struct os_mempool oc_rep_objects;
